+++ packages/typespec-client-generator-core/test/types/doc-summary.test.ts
@@ -1,28 +1,17 @@
-import { ok, strictEqual } from "assert";
-import { afterEach, beforeEach, it } from "vitest";
-import { createSdkTestRunner, SdkTestRunner } from "../test-host.js";
+import { strictEqual } from "assert";
+import { it } from "vitest";
+import { createSdkContextForTester, SimpleTesterWithService } from "../tester.js";
 
-let runner: SdkTestRunner;
-
-beforeEach(async () => {
-  runner = await createSdkTestRunner({ emitterName: "@azure-tools/typespec-java" });
-});
-afterEach(async () => {
-  for (const modelsOrEnums of [runner.context.sdkPackage.models, runner.context.sdkPackage.enums]) {
-    for (const item of modelsOrEnums) {
-      ok(item.name !== "");
-    }
-  }
-});
 it("unicode", async () => {
-  await runner.compileWithBuiltInService(`
+  const { program } = await SimpleTesterWithService.compile(`
     @usage(Usage.input | Usage.output)
     @doc("测试doc")
     @summary("测试summary")
     model Test {}
   `);
 
-  const models = runner.context.sdkPackage.models;
+  const context = await createSdkContextForTester(program);
+  const models = context.sdkPackage.models;
   strictEqual(models.length, 1);
   strictEqual(models[0].doc, "测试doc");
   strictEqual(models[0].summary, "测试summary");